match: SOURCE/ADVMGR -> 14/16 exact, and GetPlayerColor returns i32
ProcessSearch, CompleteDraw, UpdateHeroLocator, UpdateTownLocators,
HeroQuickView, TownQuickView, CastSpell, TownPortalHandler, TownGate,
ShowRoute, HideRoute, FindAdjacentMonster, APanelHandler,
UpdateSystemOptions.

CompleteDraw was a one-token semantic fix that moved a lot of bytes: retail
writes `m_mapOriginX = 0;`, not `m_mapOriginX = m_mapOriginY;`. The chained
assignment cost two instructions and rotated every register after it.

The quick-view pair needed 20- and 18-local frames plus their inner scopes, and
the same `armyStart + stride*i` add-order family at six sites.

game::GetPlayerColor returns i32, not i8. The lane pinned it on
UpdBottomViewEnemyTurn's single residual byte - a 0x66 operand-size prefix -
where retail has `movsbl 0x49c(%eax,%edx),%ecx` feeding iconWidget's short
parameter. A probe under the unit's own flags showed only an accessor declared
to return int reproduces that: the i8 accessor spills a byte temp, the direct
member read gives movsbw. Applying it here confirms the reach the lane
predicted - UpdBottomViewEnemyTurn 95.39 -> 99.98 and game::ClaimTown
87.88 -> 93.41 - with no regression anywhere.

Thirteen more English placeholders replaced with retail's CP1251 text, all
hash-verified against the image.

ComboDraw stays live at 99.31 for a reason worth recording: its 48 residual
byte-slots are one class. Retail's object carries const_00123e4f/50/51 with
zero displacement where ours carries ?bComboDraw@@3PAY0BC@CA with -0x1/+0x1/
-0x2. Same effective address, different symbol and addend split, because that
array has no DATA(...) claim - and this tree has no DATA markers at all yet.
Every other instruction in the 0x12b4-byte body is identical and the block
skeleton is 247/247.

SOURCE/ADVMGR 54/94 -> 68/94 exact; tree 1260 -> 1274, code bytes 38.91% ->
40.80%.

+++ include/SOURCE/game.h
@@ -204,7 +204,7 @@ public:
     town* GetPlayerTown(i32 player, i32 index) {
         return &m_castleRecs[m_players[player].m_townIds[index]];
     }
-    i8 GetPlayerColor(i32 player) {
+    i32 GetPlayerColor(i32 player) {
         return m_players[player].m_color;
     }
     i32 GetMineId(i32, i32);
